Deepseek Chatbot/requirements.txt
langchain
langchain-groq
python-dotenv
langchain-community
supabase